# services/backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend API",
    "version": "1.0.0",
    "description": ""
  },
  "paths": {
    "/users": {
      "post": {
        "operationId": "create_user",
        "summary": "Create User",
        "tags": [
          "users"
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "operationId": "get_user",
        "summary": "Get User",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "put": {
        "operationId": "update_user",
        "summary": "Update User",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_user",
        "summary": "Delete User",
        "tags": [
          "users"
        ],
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/debug/command": {
      "post": {
        "operationId": "command",
        "summary": "Command",
        "tags": [
          "debug"
        ],
        "responses": {
          "202": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReceived"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandReceivedCreate"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "title": "User",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "telegram_id": {
            "type": "integer",
            "minimum": 0
          },
          "is_admin": {
            "type": "boolean",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "id",
          "telegram_id",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "UserCreate": {
        "type": "object",
        "title": "UserCreate",
        "properties": {
          "telegram_id": {
            "type": "integer",
            "minimum": 0
          },
          "is_admin": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "telegram_id"
        ],
        "additionalProperties": false
      },
      "UserUpdate": {
        "type": "object",
        "title": "UserUpdate",
        "properties": {
          "telegram_id": {
            "type": "integer",
            "minimum": 0
          },
          "is_admin": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "telegram_id"
        ],
        "additionalProperties": false
      },
      "UserRead": {
        "type": "object",
        "title": "UserRead",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "telegram_id": {
            "type": "integer",
            "minimum": 0
          },
          "is_admin": {
            "type": "boolean",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "id",
          "telegram_id",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
      "UserRegisteredEvent": {
        "type": "object",
        "title": "UserRegisteredEvent",
        "properties": {
          "user_id": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "user_id",
          "email",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "CommandReceived": {
        "type": "object",
        "title": "CommandReceived",
        "properties": {
          "command": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user_id": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "command",
          "args",
          "user_id",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "CommandReceivedCreate": {
        "type": "object",
        "title": "CommandReceivedCreate",
        "properties": {
          "command": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user_id": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "command",
          "args",
          "user_id",
          "timestamp"
        ],
        "additionalProperties": false
      }
    }
  }
}