# services/backend/src/generated/routers/users.py
##############################################################################
# THIS FILE IS GENERATED BY THE FRAMEWORK. DO NOT EDIT MANUALLY.
# CHANGES WILL BE OVERWRITTEN.
##############################################################################

# AUTO-GENERATED BY SERVICE-TEMPLATE - DO NOT EDIT
"""
Router for users.
"""
# AUTO-GENERATED FROM shared/spec/rest.yaml – DO NOT EDIT MANUALLY

from __future__ import annotations

from collections.abc import Callable

from fastapi import APIRouter, Body, Depends, Path, Query  # noqa: F401
from shared.generated.schemas import (
    UserCreate,
    UserRead,
    UserUpdate,
)
from sqlalchemy.ext.asyncio import AsyncSession

from ..protocols import UsersControllerProtocol


def create_router(
    get_db: Callable[[], AsyncSession],
    get_controller: Callable[[], UsersControllerProtocol],
) -> APIRouter:
    """
    Create and configure the users router.

    Args:
        get_db: Dependency to get DB session
        get_controller: Dependency to get controller implementation

    Returns:
        Configured APIRouter instance
    """
    router = APIRouter(
        prefix="/users",
        tags=["users"],
    )

    @router.post(
        "",
        response_model=UserRead,
        status_code=201,
    )
    async def create_user(
        payload: UserCreate = Body(...),  # noqa: B008
        session: AsyncSession = Depends(get_db),  # noqa: B008
        controller: UsersControllerProtocol = Depends(get_controller),  # noqa: B008
    ) -> UserRead:
        """Handler for create_user"""
        return await controller.create_user(
            session=session,
            payload=payload,
        )

    @router.get(
        "/{user_id}",
        response_model=UserRead,
        status_code=200,
    )
    async def get_user(
        user_id: int = Path(...),
        session: AsyncSession = Depends(get_db),  # noqa: B008
        controller: UsersControllerProtocol = Depends(get_controller),  # noqa: B008
    ) -> UserRead:
        """Handler for get_user"""
        return await controller.get_user(
            session=session,
            user_id=user_id,
        )

    @router.put(
        "/{user_id}",
        response_model=UserRead,
        status_code=200,
    )
    async def update_user(
        user_id: int = Path(...),
        payload: UserUpdate = Body(...),  # noqa: B008
        session: AsyncSession = Depends(get_db),  # noqa: B008
        controller: UsersControllerProtocol = Depends(get_controller),  # noqa: B008
    ) -> UserRead:
        """Handler for update_user"""
        return await controller.update_user(
            session=session,
            user_id=user_id,
            payload=payload,
        )

    @router.delete(
        "/{user_id}",
        status_code=204,
    )
    async def delete_user(
        user_id: int = Path(...),
        session: AsyncSession = Depends(get_db),  # noqa: B008
        controller: UsersControllerProtocol = Depends(get_controller),  # noqa: B008
    ) -> None:
        """Handler for delete_user"""
        return await controller.delete_user(
            session=session,
            user_id=user_id,
        )

    return router